test: organize and cleanup tests
